make log config immutable
